+++ src/widget/create_calorie_target.rs
@@ -30,10 +30,10 @@ impl From<CreateCalorieTargetMessage> for Message {
 #[derive(Debug)]
 pub struct CreateCalorieTarget {
     day: DayFormField,
-    calories: InputFormField<f64>,
-    fats: InputFormField<f64>,
-    proteins: InputFormField<f64>,
-    carbohydrates: InputFormField<f64>,
+    calories: InputFormField<f32>,
+    fats: InputFormField<f32>,
+    proteins: InputFormField<f32>,
+    carbohydrates: InputFormField<f32>,
 }
 
 impl CreateCalorieTarget {
@@ -52,7 +52,7 @@ impl CreateCalorieTarget {
             if input.is_empty() {
                 Err(InputFormFieldError::MissingRequiredValue)
             } else {
-                match input.parse::<f64>() {
+                match input.parse::<f32>() {
                     Err(_) => Err(InputFormFieldError::InvalidNumber),
                     Ok(val) if val < 0.0 => Err(InputFormFieldError::SmallerThanZero),
                     Ok(val) => Ok(val),
@@ -64,7 +64,7 @@ impl CreateCalorieTarget {
             if input.is_empty() {
                 Err(InputFormFieldError::MissingRequiredValue)
             } else {
-                match input.parse::<f64>() {
+                match input.parse::<f32>() {
                     Err(_) => Err(InputFormFieldError::InvalidNumber),
                     Ok(val) if val < 0.0 => Err(InputFormFieldError::SmallerThanZero),
                     Ok(val) => Ok(val),
@@ -76,7 +76,7 @@ impl CreateCalorieTarget {
             if input.is_empty() {
                 Err(InputFormFieldError::MissingRequiredValue)
             } else {
-                match input.parse::<f64>() {
+                match input.parse::<f32>() {
                     Err(_) => Err(InputFormFieldError::InvalidNumber),
                     Ok(val) if val < 0.0 => Err(InputFormFieldError::SmallerThanZero),
                     Ok(val) => Ok(val),
@@ -88,7 +88,7 @@ impl CreateCalorieTarget {
             if input.is_empty() {
                 Err(InputFormFieldError::MissingRequiredValue)
             } else {
-                match input.parse::<f64>() {
+                match input.parse::<f32>() {
                     Err(_) => Err(InputFormFieldError::InvalidNumber),
                     Ok(val) if val < 0.0 => Err(InputFormFieldError::SmallerThanZero),
                     Ok(val) => Ok(val),
