+++ src/widget/update_calorie_target.rs
@@ -29,10 +29,10 @@ impl From<UpdateCalorieTargetMessage> for Message {
 #[derive(Debug)]
 pub struct UpdateCalorieTarget {
     day: NaiveDate,
-    calories: InputFormField<f64>,
-    fats: InputFormField<f64>,
-    proteins: InputFormField<f64>,
-    carbohydrates: InputFormField<f64>,
+    calories: InputFormField<f32>,
+    fats: InputFormField<f32>,
+    proteins: InputFormField<f32>,
+    carbohydrates: InputFormField<f32>,
 }
 
 impl UpdateCalorieTarget {
@@ -67,7 +67,7 @@ impl UpdateCalorieTarget {
             if input.is_empty() {
                 Err(InputFormFieldError::MissingRequiredValue)
             } else {
-                match input.parse::<f64>() {
+                match input.parse::<f32>() {
                     Err(_) => Err(InputFormFieldError::InvalidNumber),
                     Ok(val) if val < 0.0 => Err(InputFormFieldError::SmallerThanZero),
                     Ok(val) => Ok(val),
@@ -79,7 +79,7 @@ impl UpdateCalorieTarget {
             if input.is_empty() {
                 Err(InputFormFieldError::MissingRequiredValue)
             } else {
-                match input.parse::<f64>() {
+                match input.parse::<f32>() {
                     Err(_) => Err(InputFormFieldError::InvalidNumber),
                     Ok(val) if val < 0.0 => Err(InputFormFieldError::SmallerThanZero),
                     Ok(val) => Ok(val),
@@ -91,7 +91,7 @@ impl UpdateCalorieTarget {
             if input.is_empty() {
                 Err(InputFormFieldError::MissingRequiredValue)
             } else {
-                match input.parse::<f64>() {
+                match input.parse::<f32>() {
                     Err(_) => Err(InputFormFieldError::InvalidNumber),
                     Ok(val) if val < 0.0 => Err(InputFormFieldError::SmallerThanZero),
                     Ok(val) => Ok(val),
@@ -103,7 +103,7 @@ impl UpdateCalorieTarget {
             if input.is_empty() {
                 Err(InputFormFieldError::MissingRequiredValue)
             } else {
-                match input.parse::<f64>() {
+                match input.parse::<f32>() {
                     Err(_) => Err(InputFormFieldError::InvalidNumber),
                     Ok(val) if val < 0.0 => Err(InputFormFieldError::SmallerThanZero),
                     Ok(val) => Ok(val),
